bug 2006 - Exit if stdin closes, removing need for port_adaptor
This rebuilds some older functionality that was removed, but creates a
controlled shutdown centered around libevent.

stdin reading is still blocking in a separate thread (thank you, Windows).

A new periodic event is fired every one second to ensure that some
event occurs so we can shutdown the event loop.

Change-Id: I8bb6806bcc4ddf7a5602646fa359b288ccc4c932
Reviewed-on: http://review.northscale.com/1981
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>