Don't allow the stdin artificial event to block the whole system.
What I *really* want here is just alarm().  I don't get that on
Windows.  Unfortunately, I can't hop out of the event loop without
having an event to fire, so I need it there.  This change allows me to
keep the timer going until we begin a shutdown.

This does mean that we may get into situations where one channel is
shutdown, the other channel still has data to send, we aren't using
timeouts, we've stopped issuing the artificial events, and we can't
close via stdin because no events are firing.  That's a bit of a
convoluted condition documented here so I can laugh for thinking it
was so rare at some point in the future.

Change-Id: Ie7a32f6723cb221332395aaffcccfa9b8ce0f9d9
Reviewed-on: http://review.northscale.com/1989
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>