bug 2124 - unconditionally set timeouts on message pipes
There was a condition where we wouldn't set a timeout if we didn't
think we were reading from a pipe.  This left several windows where we
wouldn't have timeouts on upstream connections.

I ran this in a loop for a while and couldn't get it to happen
naturally reliably, but I simulated it by suspending the receiving
side.  The earlier code would not time out the connection and would
hang indefinintely.  This code properly times out and exits..

Change-Id: I2efb7a6134b0523d5d8548fee6edadc4408973c9
Reviewed-on: http://review.northscale.com/2044
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>